refactor(ricer::hook)!: move hook management logic into `hook` module
It makes more organizational sense to put all hook management logic into a
singular module named `ricer::hook`.
