+++ src/hook.rs
@@ -1,0 +1,316 @@
+// SPDX-FileCopyrightText: 2024 Jason Pena <jasonpena@awkless.com>
+// SPDX-License-Identifier: MIT
+
+//! Command hook management.
+//!
+//! Manage command hook definitions and hook scripts to be executed for Ricer.
+//! Command hooks only apply to Ricer's unique command set, and does not effect
+//! regular Git commands. Thus, the command hook to execute is determined by
+//! the parsed command-line arguments the user gave to Ricer, which is obtained
+//! through [`Context`].
+//!
+//! Hooks are defined in Ricer's special command hook configuration file.
+//! Commands can have multiple hook definitions stuffed into an array. Each
+//! hook definition specifies one or two hook scripts to be executed from the
+//! special `hooks/` directory. Ricer will only look for hook scripts in the
+//! `hooks/` directory.
+//!
+//! Hooks can come in two forms: _pre_ and _post_. Pre hooks are meant to be
+//! executed _before_ a given Ricer command, and post hooks are meant to execute
+//! _after_. The user can control whether or not a hook script can be executed
+//! in three ways: _always_ execute the hook no questions asked, _never_ execute
+//! the hook no questions asked, or page the hooks contents and _prompt_ the
+//! user about executing it.
+
+use crate::{
+    config::{CmdHookConfig, ConfigFile, ConfigFileError, TomlError},
+    context::{Context, HookAction},
+    locate::Locator,
+};
+
+use log::info;
+use minus::{
+    error::MinusError,
+    input::{HashedEventRegister, InputEvent},
+    page_all, ExitStrategy, LineNumbers, Pager,
+};
+use run_script::{run_script, ScriptError, ScriptOptions};
+use std::{
+    fs::read_to_string,
+    hash::RandomState,
+    io::Error as IoError,
+    path::{Path, PathBuf},
+    sync::atomic::{AtomicBool, Ordering},
+    sync::Arc,
+};
+
+#[cfg(test)]
+use mockall::automock;
+
+/// Error types for [`CmdHook`].
+#[derive(Debug, thiserror::Error)]
+pub enum CmdHookError {
+    #[error("Failed to load command hook configuration file")]
+    LoadConfig { source: ConfigFileError },
+
+    #[error("Failed to get command hook data")]
+    GetCmdHook { source: TomlError },
+
+    #[error("Failed to read hook '{path}'")]
+    HookRead { source: IoError, path: PathBuf },
+
+    #[error("Failed to run hook")]
+    RunHook { source: ScriptError },
+
+    #[error("Failed to run pager")]
+    HookPager { source: HookPagerError },
+}
+
+impl From<ConfigFileError> for CmdHookError {
+    fn from(err: ConfigFileError) -> Self {
+        CmdHookError::LoadConfig { source: err }
+    }
+}
+
+impl From<TomlError> for CmdHookError {
+    fn from(err: TomlError) -> Self {
+        CmdHookError::GetCmdHook { source: err }
+    }
+}
+
+impl From<ScriptError> for CmdHookError {
+    fn from(err: ScriptError) -> Self {
+        CmdHookError::RunHook { source: err }
+    }
+}
+
+impl From<HookPagerError> for CmdHookError {
+    fn from(err: HookPagerError) -> Self {
+        CmdHookError::HookPager { source: err }
+    }
+}
+
+/// Error types for [`HookPager`].
+#[derive(Debug, thiserror::Error)]
+pub enum HookPagerError {
+    #[error("Minus pager failed because '{source}'")]
+    Minus { source: MinusError },
+}
+
+impl From<MinusError> for HookPagerError {
+    fn from(err: MinusError) -> Self {
+        HookPagerError::Minus { source: err }
+    }
+}
+
+/// Command hook execution handler.
+///
+/// Executes user-defined hooks according to specified hook actions selected by
+/// the user through [`Context`]. Hooks can be defined as _pre_ or _post_, i.e.,
+/// run _before_ a command, or run _after_ a command. Hooks utilize scripts that
+/// must be defined in the special `hooks/` directory at the top-level of
+/// Ricer's configuration directory.
+///
+/// User can set hook actions to _never_, _always_, and _prompt_. Never action
+/// means that no hook can be executed no questions asked. Always action means
+/// that hooks are executed no questions asked. Finally, prompt action will page
+/// the contents of a hook script for the user to review, and prompt them about
+/// whether or not they want to execute it.
+pub struct CmdHook<'cfg, L>
+where
+    L: Locator,
+{
+    context: &'cfg Context,
+    locator: &'cfg L,
+    config: ConfigFile<'cfg, CmdHookConfig, L>,
+    pager: HookPager,
+}
+
+impl<'cfg, L> CmdHook<'cfg, L>
+where
+    L: Locator,
+{
+    /// Load new command hook handler.
+    ///
+    /// Will load the contents of the command hook configuration file based
+    /// on the path provided by `locator`. Will also load user selected actions
+    /// from `context`.
+    ///
+    /// # Errors
+    ///
+    /// 1. Return [`CmdHookError::LoadConfig`] if configuration file cannot be
+    ///    read and parsed for some reason.
+    ///
+    /// # See also
+    ///
+    /// - [`ConfigFile`]
+    /// - [`Locator`]
+    pub fn load(context: &'cfg Context, locator: &'cfg L) -> Result<Self, CmdHookError> {
+        let config = ConfigFile::load(CmdHookConfig, locator)?;
+        Ok(Self { context, locator, config, pager: Default::default() })
+    }
+
+    pub fn set_pager(&mut self, pager: HookPager) {
+        self.pager = pager;
+    }
+
+    /// Run user-defined hooks.
+    ///
+    /// Run specific hook kind for given command that was selected through
+    /// [`Context`].
+    ///
+    /// # Errors
+    ///
+    /// 1. Return [`CmdHookError::GetCmdHook`] if current command hook
+    ///    definition cannot be obtained through hook configuration file.
+    /// 2. Return [`CmdHookError::HookRead`] if hook script cannot be read
+    ///    from `hooks/` directory.
+    /// 3. Return [`CmdHookError::RunHook`] if hook script cannot be executed
+    ///    for whatever reason.
+    /// 4. Return [`CmdHookError::HookPager`] if pager cannot page hook script
+    ///    and prompt user.
+    pub fn run_hooks(&self, hook_kind: HookKind) -> Result<(), CmdHookError> {
+        // INVARIANT: Git command shortcut cannot execute hooks.
+        if matches!(self.context, Context::Git(..)) {
+            return Ok(());
+        }
+
+        let action = self.get_hook_action();
+        if action == &HookAction::Never {
+            return Ok(());
+        }
+
+        let cmd_hook = self.config.get(self.context.to_string())?;
+        for hook in cmd_hook.hooks {
+            let hook_name = match hook_kind {
+                HookKind::Pre => hook.pre.as_ref(),
+                HookKind::Post => hook.post.as_ref(),
+            };
+            let hook_name = match hook_name {
+                Some(name) => name,
+                None => continue, // Skip this iteration if no hook name is found.
+            };
+
+            let hook_path = self.locator.hooks_dir().join(hook_name);
+            let hook_data = read_to_string(&hook_path)
+                .map_err(|err| CmdHookError::HookRead { source: err, path: hook_path.clone() })?;
+
+            if action == &HookAction::Prompt {
+                let workdir = hook.workdir.as_deref().unwrap_or(self.locator.hooks_dir());
+                self.pager.page_and_prompt(hook_path.as_path(), workdir, &hook_data)?;
+                if !self.pager.choice() {
+                    continue; // Skip this iteration if user denied hook script.
+                }
+            }
+
+            let mut hook_opts = ScriptOptions::new();
+            hook_opts.working_directory = hook.workdir;
+            let (code, out, err) = run_script!(hook_data, hook_opts)?;
+            info!("({code}) {}\nstdout: {out}\nstderr: {err}", hook_path.display());
+        }
+
+        Ok(())
+    }
+
+    fn get_hook_action(&self) -> &HookAction {
+        match self.context {
+            Context::Bootstrap(ctx) => &ctx.shared.run_hook,
+            Context::Clone(ctx) => &ctx.shared.run_hook,
+            Context::Commit(ctx) => &ctx.shared.run_hook,
+            Context::Delete(ctx) => &ctx.shared.run_hook,
+            Context::Enter(ctx) => &ctx.shared.run_hook,
+            Context::Init(ctx) => &ctx.shared.run_hook,
+            Context::List(ctx) => &ctx.shared.run_hook,
+            Context::Pull(ctx) => &ctx.shared.run_hook,
+            Context::Push(ctx) => &ctx.shared.run_hook,
+            Context::Rename(ctx) => &ctx.shared.run_hook,
+            Context::Status(ctx) => &ctx.shared.run_hook,
+
+            // INVARIANT: Git command shortcut cannot use hooks.
+            Context::Git(_) => {
+                unreachable!("This should not happen. Git shortcut cannot use hooks")
+            }
+        }
+    }
+}
+
+/// Hook type to execute.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum HookKind {
+    /// Execute hooks _before_ command.
+    Pre,
+
+    /// Execute hooks _after_ command.
+    Post,
+}
+
+/// Pager for hook scripts.
+///
+/// Basic static pager that shows the current contents of a given hook script,
+/// and prompts the user about whether or not they want to execute it. User
+/// can accept or deny hook script by pressing "a" or "d".
+///
+/// # See also
+///
+/// - [Minus](https://docs.rs/minus/latest/minus/)
+#[derive(Default)]
+pub struct HookPager {
+    choice: Arc<AtomicBool>,
+}
+
+#[cfg_attr(test, automock)]
+impl HookPager {
+    pub fn new() -> Self {
+        Self { choice: Arc::new(AtomicBool::default()) }
+    }
+
+    pub fn choice(&self) -> bool {
+        self.choice.load(Ordering::Relaxed)
+    }
+
+    /// Page hook script and prompt user about running it.
+    ///
+    /// # Errors
+    ///
+    /// - Return [`HookPagerError::Minus`] for any issues encountered with
+    ///   [Minus](https://docs.rs/minus/latest/minus/).
+    pub fn page_and_prompt(
+        &self,
+        file_name: &Path,
+        workdir: &Path,
+        file_data: &str,
+    ) -> Result<(), HookPagerError> {
+        let pager = Pager::new();
+        pager.set_prompt(format!(
+            "Run '{}' at '{}'? [a]ccept/[d]eny",
+            file_name.display(),
+            workdir.display(),
+        ))?;
+        pager.show_prompt(true)?;
+        pager.set_run_no_overflow(true)?;
+        pager.set_line_numbers(LineNumbers::Enabled)?;
+        pager.push_str(file_data)?;
+        pager.set_input_classifier(self.generate_key_bindings())?;
+        pager.set_exit_strategy(ExitStrategy::PagerQuit)?;
+        page_all(pager)?;
+        Ok(())
+    }
+
+    fn generate_key_bindings(&self) -> Box<HashedEventRegister<RandomState>> {
+        let mut input = HashedEventRegister::default();
+
+        let response = self.choice.clone();
+        input.add_key_events(&["a"], move |_, _| {
+            response.store(true, Ordering::Relaxed);
+            InputEvent::Exit
+        });
+
+        let response = self.choice.clone();
+        input.add_key_events(&["d"], move |_, _| {
+            response.store(false, Ordering::Relaxed);
+            InputEvent::Exit
+        });
+
+        Box::new(input)
+    }
+}
